requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
torch==2.7.0
torchvision==0.22.0
pillow==11.2.1
requests==2.32.3
gunicorn==22.0.0